requirements.txt
fastapi~=0.115.8
psycopg2-binary
python-dotenv~=1.0.1
fastapi[all]
uvicorn~=0.34.0
asyncpg
fastapi-cache2[redis]
redis~=5.2.1
aioredis
gunicorn
pydantic~=2.10.6
starlette~=0.45.3
python-jose[cryptography]
passlib
peewee
Jinja2
pytest
pytest-cov
pytest-asyncio
coverage
requests
requests-mock
bcrypt==4.0.1
locust